+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v2.0.0-beta8 - 2023-05-xx
 - Add --connect=<value> option to start a stream op startup #805 #509 
+- Add video zoom option (Credits to @adolmarro) #667
 
 ## v2.0.0-beta7 - 2023-04-11
 - Smoothen login experience by saving the login tokens and attempt authentication before popup prompt #782